+++ R/knitr.R
@@ -38,11 +38,15 @@ knit_print.huxtable <- function(x, options, ...) {
     }
   )
 
-  res <- do.call(call_name, list(x))
+  call_args <- list(x)
+  if (of %in% c("html", "latex")) call_args$dependencies <- FALSE
+  res <- do.call(call_name, call_args)
 
   res <- switch(of,
     latex = {
       latex_deps <- report_latex_dependencies(quiet = TRUE)
+      array_dep <- match("array", vapply(latex_deps, `[[`, "", "name"))
+      latex_deps[[array_dep]]$extra_lines <- huxtable_latex_commands()
       tenv <- tabular_environment(x)
       if (tenv %in% "tabulary") {
         latex_deps <- c(latex_deps, list(rmarkdown::latex_dependency(tenv)))
@@ -50,7 +54,8 @@ knit_print.huxtable <- function(x, options, ...) {
       knitr::asis_output(res, meta = latex_deps)
     },
     html = knitr::asis_output(
-      htmltools::htmlPreserve(res)
+      htmltools::htmlPreserve(res),
+      meta = list(huxtable_html_dependency())
     ),
     rtf = knitr::raw_output(res),
     pptx = ,