+++ R/latex.R
@@ -9,16 +9,15 @@ default_table_width_unit <- "\\textwidth"
 
 #' @export
 #' @rdname to_latex
-print_latex <- function(ht, ...) {
-  cat(to_latex(ht, ...))
+print_latex <- function(ht, dependencies = TRUE, ...) {
+  cat(to_latex(ht, ..., dependencies = dependencies))
 }
 
 
 #' Create LaTeX representing a huxtable
 #'
-#' @param ht A huxtable.
+#' @inheritParams to_html
 #' @param tabular_only Return only the LaTeX tabular, not the surrounding float.
-#' @param ... Arguments passed to methods.
 #'
 #' @details
 #' If we appear to be in a rmarkdown document with the Pandoc markdown `+raw_attribute` extension
@@ -37,8 +36,8 @@ print_latex <- function(ht, ...) {
 #'   b = letters[1:3]
 #' )
 #' print_latex(ht)
-to_latex <- function(ht, tabular_only = FALSE, ...) {
-  assert_that(is.flag(tabular_only))
+to_latex <- function(ht, tabular_only = FALSE, dependencies = TRUE, ...) {
+  assert_that(is.flag(tabular_only), is.flag(dependencies))
   if (breakable(ht)) {
     if (!is.na(height(ht))) {
       stop("Breakable LaTeX tables cannot have a fixed height.", call. = FALSE)
@@ -57,11 +56,11 @@ to_latex <- function(ht, tabular_only = FALSE, ...) {
       "\\colorbox[RGB]{", bg, "}{", tabular, "}}"
     )
   }
-  commands <- "
-  \\providecommand{\\huxb}[2]{\\arrayrulecolor[RGB]{#1}\\global\\arrayrulewidth=#2pt}
-  \\providecommand{\\huxvb}[2]{\\color[RGB]{#1}\\vrule width #2pt}
-  \\providecommand{\\huxtpad}[1]{\\rule{0pt}{#1}}
-  \\providecommand{\\huxbpad}[1]{\\rule[-#1]{0pt}{#1}}\n"
+  commands <- if (dependencies) {
+    paste0("\n  ", paste(huxtable_latex_commands(), collapse = "\n  "), "\n")
+  } else {
+    ""
+  }
 
   if (tabular_only) {
     return(maybe_markdown_fence(paste0(commands, tabular)))