+++ tests/testthat/test-latex-dependencies.R
@@ -21,6 +21,7 @@ test_that("install/report_latex_dependencies", {
     as_string = TRUE
   ))
   expect_match(package_str, "\\\\usepackage\\{array\\}")
+  expect_no_match(package_str, "\\\\providecommand\\{\\\\huxb\\}")
 })